+++ src/index.ts
@@ -2,9 +2,12 @@ export {
   Button,
   type ButtonProps,
   TheMusicTreeByline,
+  TheMusicTreeHorizontalLink,
   getOrgSiteHref,
   ORG_URL,
+  parseOrgSiteHref,
   type TheMusicTreeBylineProps,
+  type TheMusicTreeHorizontalLinkProps,
 } from "./components";
 export * from "./tokens";
 export * from "./hooks";